a lot

+++ route_generator.py
@@ -12,14 +12,12 @@ import math
 from typing import List, Tuple
 from heapq import heappop, heappush
 
-
 # euclidean_distance function is used to calculate the Euclidean distance between two coordinates
-def euclidean_distance(coord1: Tuple[int, int], coord2: Tuple[int, int]) -> float:
+def euclidean_distance(coord1: Tuple[float, float], coord2: Tuple[float, float]) -> float:
     return math.sqrt((coord1[0] - coord2[0]) ** 2 + (coord1[1] - coord2[1]) ** 2)
 
-
 # construct_mst function is used to construct the Minimum Spanning Tree (MST) of the given coordinates
-def construct_mst(coordinates: List[Tuple[int, int]]) -> List[List[int]]:
+def construct_mst(coordinates: List[Tuple[float, float]]) -> List[List[int]]:
     n = len(coordinates)
     adj_list = [[] for _ in range(n)]
     visited = [False] * n
@@ -41,7 +39,6 @@ def construct_mst(coordinates: List[Tuple[int, int]]) -> List[List[int]]:
 
     return adj_list
 
-
 # dfs_traversal function is used to perform Depth First Search (DFS) traversal on the MST
 def dfs_traversal(adj_list: List[List[int]], start: int = 0) -> List[int]:
     visited = [False] * len(adj_list)
@@ -57,19 +54,16 @@ def dfs_traversal(adj_list: List[List[int]], start: int = 0) -> List[int]:
     dfs(start)
     return route
 
-
 # generate_route function is used to generate the route based on the DFS traversal
-def generate_route(coordinates: List[Tuple[int, int]]) -> List[int]:
+def generate_route(coordinates: List[Tuple[float, float]]) -> List[int]:
     mst = construct_mst(coordinates) 
     route = dfs_traversal(mst)       
     route.append(route[0])           
     return route
 
-
 # calculate_total_distance function is used to calculate the total distance of the generated route
-def calculate_total_distance(route: List[int], coordinates: List[Tuple[int, int]]) -> float:
+def calculate_total_distance(route: List[int], coordinates: List[Tuple[float, float]]) -> float:
     total_distance = 0
     for i in range(len(route) - 1):
         total_distance += euclidean_distance(coordinates[route[i]], coordinates[route[i + 1]])
     return total_distance
-
